requirements.txt
asgiref==3.5.2
Django==4.0.4
django-ckeditor==6.3.2
django-cleanup==6.0.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
